feat: add support of exporting confusion matrix for 3D analyzer
