+++ perception_eval/test/perception_lsim.py
@@ -241,11 +241,13 @@ if __name__ == "__main__":
     # Detection performance report
     detection_analyzer = PerceptionAnalyzer3D(detection_lsim.evaluator.evaluator_config)
     detection_analyzer.add(detection_lsim.evaluator.frame_results)
-    score_df, error_df = detection_analyzer.analyze()
+    score_df, error_df, conf_mat_df = detection_analyzer.analyze()
     if score_df is not None:
         logging.info(score_df.to_string())
     if error_df is not None:
         logging.info(error_df.to_string())
+    if conf_mat_df is not None:
+        logging.info(conf_mat_df.to_string())
 
     # detection_analyzer.plot_state("4bae7e75c7de70be980ce20ce8cbb642", ["x", "y"])
     # detection_analyzer.plot_error(["x", "y"])
@@ -320,11 +322,13 @@ if __name__ == "__main__":
     # Tracking performance report
     tracking_analyzer = PerceptionAnalyzer3D(tracking_lsim.evaluator.evaluator_config)
     tracking_analyzer.add(tracking_lsim.evaluator.frame_results)
-    score_df, error_df = tracking_analyzer.analyze()
+    score_df, error_df, conf_mat_df = tracking_analyzer.analyze()
     if score_df is not None:
         logging.info(score_df.to_string())
     if error_df is not None:
         logging.info(error_df.to_string())
+    if conf_mat_df is not None:
+        logging.info(conf_mat_df.to_string())
 
     # Clean up tmpdir
     if args.use_tmpdir:
