+++ perception_eval/test/perception_lsim2d.py
@@ -59,9 +59,11 @@ class PerceptionLSimMoc:
         # If target_labels = None, all labels will be evaluated.
         evaluation_config_dict.update(
             dict(
-                target_labels=["green", "red", "yellow", "unknown"]
-                if label_prefix == "traffic_light"
-                else ["car", "bicycle", "pedestrian", "motorbike"],
+                target_labels=(
+                    ["green", "red", "yellow", "unknown"]
+                    if label_prefix == "traffic_light"
+                    else ["car", "bicycle", "pedestrian", "motorbike"]
+                ),
                 ignore_attributes=["cycle_state.without_rider"] if label_prefix == "autoware" else None,
             )
         )
@@ -242,9 +244,11 @@ if __name__ == "__main__":
     # Detection performance report
     detection_analyzer = PerceptionAnalyzer2D(detection_lsim.evaluator.evaluator_config)
     detection_analyzer.add(detection_lsim.evaluator.frame_results)
-    score_df, conf_mat_df = detection_analyzer.analyze()
+    score_df, error_df, conf_mat_df = detection_analyzer.analyze()
     if score_df is not None:
         logging.info(score_df.to_string())
+    if error_df is not None:
+        logging.info(error_df.to_string())
     if conf_mat_df is not None:
         logging.info(conf_mat_df.to_string())
 
@@ -278,9 +282,11 @@ if __name__ == "__main__":
     # Tracking performance report
     tracking_analyzer = PerceptionAnalyzer2D(tracking_lsim.evaluator.evaluator_config)
     tracking_analyzer.add(tracking_lsim.evaluator.frame_results)
-    score_df, conf_mat_df = tracking_analyzer.analyze()
+    score_df, error_df, conf_mat_df = tracking_analyzer.analyze()
     if score_df is not None:
         logging.info(score_df.to_string())
+    if error_df is not None:
+        logging.info(error_df.to_string())
     if conf_mat_df is not None:
         logging.info(conf_mat_df.to_string())
 
@@ -310,7 +316,7 @@ if __name__ == "__main__":
     # Classification performance report
     classification_analyzer = PerceptionAnalyzer2D(classification_lsim.evaluator.evaluator_config)
     classification_analyzer.add(classification_lsim.evaluator.frame_results)
-    score_df, conf_mat_df = classification_analyzer.analyze()
+    score_df, _, conf_mat_df = classification_analyzer.analyze()
     if score_df is not None:
         logging.info(score_df.to_string())
     if conf_mat_df is not None:
